[login, signup]move/verifyAuth, controlUI 파일 분리
