+++ js/verifyAuth.js
@@ -1,33 +1,4 @@
-//에러 메시지 표시 함수
-export const showError = (input, msg) => {
-    clearError(input);
-
-    input.classList.add("show-error");
-
-    const errorMsg = document.createElement("div");
-    errorMsg.classList.add("error-msg");
-    errorMsg.innerText = msg;
-
-    input.parentElement.insertAdjacentElement("afterend", errorMsg);
-}
-
-//에러 메시지 제거 함수
-export const clearError = (input) => {
-    const errorMsg = input.parentElement.nextElementSibling;
-    input.classList.remove("show-error");
-    if (errorMsg && errorMsg.classList.contains("error-msg")) {
-        errorMsg.remove()
-    }
-};
-
-//모달 팝업 표시 함수
-export const showModal = (msg, modal, modalMsg) => {
-    modalMsg.innerHTML = msg;
-    modal.classList.add("show")
-}
-
-//모달 팝업 제거 함수
-export const closeModal = (modal) => modal.classList.remove("show");
+import { showError, clearError } from "./controlUI.js";
 
 //이메일 유효성 검사 함수
 export const validEmail = (e, msgList, emailRegex, toggleBtnState) => {
@@ -71,12 +42,33 @@ export const validPw = (e, msgList, pwRegex, toggleBtnState) => {
     toggleBtnState();
 }
 
-//비밀번호 보기 함수
-export const visibleOn = (e) => {
-    const btn = e.target;
-    const input = btn.previousElementSibling;
-    input.classList.toggle("active-visible");
+//닉네임 확인 함수
+export const validNickname = (e, msgList, toggleBtnState) => {
+    const nickname = e.target;
 
-    input.type = input.classList.contains("active-visible") ? "text" : "password";
-    btn.src = input.classList.contains("active-visible") ? "./assets/btn_visibility_on.svg" : "./assets/btn_visibility_off.svg";
+    if (!nickname.value.trim()) {
+        showError(nickname, msgList.nickameInputErrorMsg);
+    }
+    else {
+        clearError(nickname);
+    }
+
+    toggleBtnState();
+}
+
+//비밀번호 확인 함수
+export const confirmPw = (e, msgList, pwInput, toggleBtnState) => {
+    const targetPw = e.target;
+
+    if (!targetPw.value.trim()) {
+        showError(targetPw, msgList.pwInputErrorMsg);
+    }
+    else if (targetPw.value !== pwInput.value) {
+        showError(targetPw, msgList.pwConfirmMsg);
+    }
+    else {
+        clearError(targetPw);
+    }
+
+    toggleBtnState();
 }
