+++ login.html
@@ -84,6 +84,7 @@
     </main>
 
     <script type="module" src="./js/verifyAuth.js"></script>
+    <script type="module" src="./js/controlUI.js"></script>
     <script type="module" src="./js/login.js"></script>
 </body>
 </html>
