+++ signup.html
@@ -103,6 +103,7 @@
     </main>
 
     <script type="module" src="./js/verifyAuth.js"></script>
+    <script type="module" src="./js/controlUI.js"></script>
     <script type="module" src="./js/signup.js"></script>
 </body>
 </html>
